learning-plan-recommender/requirements.txt
Flask==2.0.1
openai==0.11.3
python-dotenv==0.19.1
requests==2.26.0
pytest==6.2.5